Testing updating watchdog and timing
The board is not crashing when the pulse is not intermediately  triggered, but once it is (asynch), no pulse is shown in the output and the level of the last window is ~1500, when the rest of the windows are ~1200. Then, the dig time is 7 us and the output is  similar to asynchronous pulse even when is synchronous.